% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/launch_route-explorer.R
\name{launch_route_explorer}
\alias{launch_route_explorer}
\title{Launch route explorer shiny}
\usage{
launch_route_explorer()
}
\value{
A shiny app is launched
}
\description{
Launch route explorer shiny
}
\examples{
\dontrun{
launch_route_explorer()
}
}
\seealso{
https://stackoverflow.com/a/49623819/16502170
}
