% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_gpx.R
\name{load_gpx}
\alias{load_gpx}
\title{Load .gpx files from disk}
\usage{
load_gpx(path)
}
\arguments{
\item{path}{A character vector with one element. This should be the path to a locally stored gpx file.}
}
\value{
A simple features (sf) dataframe
}
\description{
This function loads a locally stored .gpx file and returns a simple features (sf) dataframe.
}
\examples{
\dontrun{   
activity1 <- load_gpx("C:/Users/my_user_name/activities/activity1.gpx")
}

}
