% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process_strava_archive.R
\name{process_strava_archive}
\alias{process_strava_archive}
\title{Unzip a Strava archive and return a dataframe of activities}
\usage{
process_strava_archive(zip_file_path, outdir = getwd())
}
\arguments{
\item{zip_file_path}{A single character vector. It should be the path to where the archive is stored on your computer}

\item{outdir}{A single character vector.
This should be the file path of the folder where you want to store the unzipped files. Defaults to the current working directory.}
}
\value{
A dataframe of activities
}
\description{
This function unzips a strava archive stored in a zip file.
The zip file is loaded from the path at \code{zip_file_path}.
The  unzipped file are stored in \code{outdir}. If outdir doesn't already exist, it will be created. Multiple folders can be created recursively.
The function returns a dataframe summarizing the activities stored in the archive.
See Strava's guide to downloading your archive: https://support.strava.com/hc/en-us/articles/216918437-Exporting-your-Data-and-Bulk-Export#h_01GG58HC4F1BGQ9PQZZVANN6WF

Note: Use this function for bulk downloads not individual activities!
}
\examples{
# Return a dataframe of activies
activities_df <- process_strava_archive(zip_file_path = "C:/Users/my_user_name/downloads/export_1111111.zip", 
                                        outdir = "./data")

# unzip the archive without creating a dataframe of activies
# return files to the current working directory

process_strava_archive(zip_file_path = "C:/Users/my_user_name/downloads/export_1111111.zip")

}
