% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_gpx.R
\name{load_gpx}
\alias{load_gpx}
\title{Title}
\usage{
load_gpx(path)
}
\arguments{
\item{path}{A character vector with one element. This should be the path to a locally stored gpx file.}
}
\value{
A dataframe
}
\description{
This function allows you to vectorise multiple \code{\link[=switch]{switch()}} statements. Each
case is evaluated sequentially and the first match for each element
determines the corresponding value in the output vector. If no cases match,
the \code{.default} is used.
}
