% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_gpx.R
\name{load_gpx}
\alias{load_gpx}
\title{Load a .gpx file from disk}
\usage{
load_gpx(path)
}
\arguments{
\item{path}{A character vector with one element. This should be the path to a locally stored gpx file.}
}
\value{
A dataframe
}
\description{
This function loads a locally stored .gpx file using the track_points layer
and returns a simple features (sf) data frame
}
\examples{
activity1 <- load_gpx("C:/Users/your user name/activities/activity1.gpx")
}
